FEATURE: ONE-4071 Change behavior of the Exclude this period date filter
Based on discussion with Jakub Sterba, the  `Exclude this period`
checkbox should move the whole interval, not just exclude last period.

The filter titles may not be ideal now (for example "From 7 to 1 months
ago" for the "Last half year" filter), but we decided to keep it that
way for now.